requirements.txt
streamlit
pandas
duckdb
pyyaml
